+++ src/main/java/domain/betting/ParticipantsProfitResult.java
@@ -1,6 +1,7 @@
 package domain.betting;
 
 import java.util.List;
+import view.dto.PlayerProfitResult;
 
 public record ParticipantsProfitResult(
         int dealerProfit,
