+++ src/main/java/view/dto/PlayerProfitResult.java
@@ -1,4 +1,4 @@
-package domain.betting;
+package view.dto;
 
 public record PlayerProfitResult(
         String name,
