+++ src/test/java/domain/betting/BettingAmountParserTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import view.BettingAmountParser;
 
 class BettingAmountParserTest {
     @ParameterizedTest
